refactor(entity): 将 Friendship 类中的 Long 类型改为 Integer 类型
- 将 id、userId 和 friendId 字段的类型从 Long 改为 Integer
- 这一变更可以减少内存占用，并且对于社交网络应用来说，Integer 类型已经足够使用

+++ backend/src/main/java/cn/hmck/entity/Friendship.java
@@ -11,11 +11,11 @@ import java.time.LocalDateTime;
 @TableName("er_friendship")
 public class Friendship implements Serializable {
     @TableId(type = IdType.AUTO)
-    private Long id;  // 朋友关系ID
+    private Integer id;  // 朋友关系ID
 
-    private Long userId;  // 用户ID
+    private Integer userId;  // 用户ID
 
-    private Long friendId;  // 朋友ID
+    private Integer friendId;  // 朋友ID
 
     @TableField(fill = FieldFill.INSERT)
     private LocalDateTime createdAt;  // 创建时间
